+++ backend/src/Docs/Hasql/Sessions.hs
@@ -62,7 +62,7 @@ import Docs.TreeRevision
     , TreeRevisionRef (..)
     )
 import DocumentManagement.Hash (Hash)
-import GHC.Int (Int32)
+import GHC.Int (Int64)
 
 existsDocument :: DocumentID -> Session Bool
 existsDocument = flip statement Statements.existsDocument
@@ -146,19 +146,19 @@ getTree rootHash = do
             (TreeEdgeToNode hash header) -> fromHeader hash header <&> Tree.Tree
 
 getTextRevisionHistory
-    :: TextElementRef -> Maybe UTCTime -> Int32 -> Session TextRevisionHistory
+    :: TextElementRef -> Maybe UTCTime -> Int64 -> Session TextRevisionHistory
 getTextRevisionHistory ref before limit =
     statement (ref, before, limit) Statements.getTextRevisionHistory
         <&> TextRevisionHistory ref . Vector.toList
 
 getTreeRevisionHistory
-    :: DocumentID -> Maybe UTCTime -> Int32 -> Session TreeRevisionHistory
+    :: DocumentID -> Maybe UTCTime -> Int64 -> Session TreeRevisionHistory
 getTreeRevisionHistory id_ before limit =
     statement (id_, before, limit) Statements.getTreeRevisionHistory
         <&> TreeRevisionHistory id_ . Vector.toList
 
 getDocumentRevisionHistory
-    :: DocumentID -> Maybe UTCTime -> Int32 -> Session DocumentHistory
+    :: DocumentID -> Maybe UTCTime -> Int64 -> Session DocumentHistory
 getDocumentRevisionHistory id_ before limit =
     statement (id_, before, limit) Statements.getDocumentRevisionHistory
         <&> DocumentHistory id_ . Vector.toList
